#39 refactored models to single file
