+++ app/controllers/auth_controller.py
@@ -1,6 +1,6 @@
 from typing import Tuple
 from flask import request, Blueprint, render_template
-from app.models.user import User
+from app.models import User
 from app.services import auth_service
 from app.utils.validation import is_email_valid
 from app.utils.session_decorators import session_required
