--- app/models/post.py
@@ -1,15 +1,0 @@
-from app.extensions import db
-
-
-class Post(db.Model):
-
-    __tablename__ = 'post'
-
-    post_id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.user_id'), nullable=False)
-    image_url = db.Column(db.String(255), nullable=False)
-    description = db.Column(db.String(255), nullable=False)
-    created_at = db.Column(db.DateTime)
-
-    def __repr__(self):
-        return f'<Post post_id={self.post_id}>'
